Move config wizard system forms to system directory (#1097)
* Move system forms to system directory

* Refactor form layout components into its own file

* Rename files from form --> step

* Update changelog
