+++ clients/ctl/admin-ui/src/features/system/PrivacyDeclarationAccordion.tsx
@@ -5,30 +5,15 @@ import {
   AccordionItem,
   AccordionPanel,
   Box,
-  Grid,
-  GridItem,
   Stack,
   Text,
 } from "@fidesui/react";
-import { ReactNode } from "react";
 
 import TaxonomyEntityTag from "~/features/taxonomy/TaxonomyEntityTag";
 import { PrivacyDeclaration } from "~/types/api";
 
-const DeclarationItem = ({
-  label,
-  children,
-}: {
-  label: string;
-  children: ReactNode;
-}) => (
-  <Grid templateColumns="1fr 2fr" data-testid={`declaration-${label}`}>
-    <GridItem>
-      <Text color="gray.600">{label}</Text>
-    </GridItem>
-    <GridItem>{children}</GridItem>
-  </Grid>
-);
+import { DeclarationItem } from "./form-layout";
+
 interface Props {
   privacyDeclaration: PrivacyDeclaration;
 }
