+++ clients/ctl/admin-ui/src/features/system/SystemRegisterSuccess.tsx
